v0.2.0 - LangInterface trait change
Earlier LangInterface::print_line used to take only the line to print,
now it takes all the lines and the index of which line to print.
This caused issues while linting multi-line programs which had specific
things over multiple lines. For example take the following lines:
```
<some-code> /* Multi
    line
Comment */
```
This probably would not be linted properly, since in line 2, the context
that the word 'line' is within a comment is lost.